+++ docs/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.32.0 (unreleased to Maven Central)
+
+- Fixes https://github.com/structurizr/dsl/issues/308 (Hidden (e.g. .DS_Store) file causes exception during !include <directory>).
+
 ## 1.31.0 (25th July 2023)
 
 - Adds support for passing parameters to external scripts.